requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
scipy
numpy
pyyaml
tqdm
sentencepiece
protobuf

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
